refactor4
