+++ bin/debug-trace-server/Cargo.toml
@@ -32,6 +32,7 @@ revm-inspectors.workspace = true
 # stateless
 stateless-common = { path = "../../crates/stateless-common" }
 stateless-core = { path = "../../crates/stateless-core" }
+stateless-db = { path = "../../crates/stateless-db" }
 
 # misc
 clap.workspace = true
