+++ crates/stateless-core/Cargo.toml
@@ -40,10 +40,12 @@ reth-trie-sparse.workspace = true
 revm.workspace = true
 
 # misc
+bincode.workspace = true
 dyn-clone.workspace = true
 eyre.workspace = true
 kanal.workspace = true
 num_cpus.workspace = true
+redb.workspace = true
 rustc-hash.workspace = true
 serde.workspace = true
 serde_json.workspace = true
@@ -53,9 +55,6 @@ tokio-util.workspace = true
 tracing.workspace = true
 
 [dev-dependencies]
-# misc
-bincode.workspace = true
-
 # stateless
 stateless-test-utils = { path = "../stateless-test-utils" }
