+++ crates/stateless-core/src/pipeline/config.rs
@@ -1,0 +1,80 @@
+//! Pipeline configuration and outcome types.
+
+use std::time::Duration;
+
+use alloy_primitives::{BlockHash, BlockNumber};
+
+/// Configuration for the chain sync pipeline.
+///
+/// Binary-specific settings (metrics, reporting, pruner) live in binary CLI args.
+#[derive(Debug, Clone)]
+pub struct PipelineConfig {
+    /// Number of parallel processing workers.
+    pub concurrent_workers: usize,
+    /// Optional block height to sync to; `None` for infinite sync.
+    pub sync_target: Option<u64>,
+    /// Time to wait between fetcher poll cycles when ahead of chain.
+    pub poll_interval: Duration,
+    /// Time to wait when pipeline encounters errors before restarting.
+    pub error_restart_delay: Duration,
+    /// Channel capacity for the fetch→worker pipeline.
+    pub fetch_channel_capacity: usize,
+    /// Channel capacity for the worker→advancer pipeline.
+    pub result_channel_capacity: usize,
+    /// Maximum number of in-flight block fetches.
+    pub fetcher_max_in_flight: usize,
+    /// Maximum RPC retry backoff for the fetcher.
+    pub fetcher_max_backoff: Duration,
+    /// If local tip falls behind remote by more than this, reset anchor.
+    /// `None` = disabled (validator). `Some(N)` = enabled (trace server).
+    pub stale_reset_threshold: Option<u64>,
+}
+
+impl Default for PipelineConfig {
+    fn default() -> Self {
+        // Physical cores: workers are CPU-bound (EVM + IPA), hyperthreads don't help.
+        let workers = num_cpus::get_physical();
+        Self {
+            concurrent_workers: workers,
+            sync_target: None,
+            poll_interval: Duration::from_millis(100),
+            error_restart_delay: Duration::from_secs(1),
+            fetch_channel_capacity: 2 * workers,
+            result_channel_capacity: 2 * workers,
+            fetcher_max_in_flight: workers,
+            fetcher_max_backoff: Duration::from_secs(30),
+            stale_reset_threshold: None,
+        }
+    }
+}
+
+/// Outcome of a single pipeline cycle.
+#[derive(Debug)]
+pub enum PipelineOutcome {
+    /// Clean shutdown (cancellation token fired or channel closed).
+    Shutdown,
+    /// Reorg detected — caller should rollback and restart.
+    Reorg(ReorgEvent),
+    /// Deterministic failure (e.g., validation mismatch) — halt, no point retrying.
+    Fatal(String),
+}
+
+/// Whether a processing error is worth retrying.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ErrorAction {
+    /// Transient error (RPC timeout, network issue) — restart pipeline, retry later.
+    Retry,
+    /// Deterministic failure (validation mismatch) — halt immediately.
+    Halt,
+}
+
+/// Details of a detected chain reorganization.
+#[derive(Debug)]
+pub struct ReorgEvent {
+    /// Block number to roll back to (inclusive — this block stays).
+    pub rollback_to: BlockNumber,
+    /// Number of blocks being reverted.
+    pub depth: u64,
+    /// Hashes of blocks being reverted (for cache invalidation).
+    pub reverted_hashes: Vec<BlockHash>,
+}
