+++ crates/stateless-db/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "stateless-db"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+exclude.workspace = true
+
+[dependencies]
+# alloy
+alloy-primitives.workspace = true
+alloy-rpc-types-eth.workspace = true
+
+# op
+op-alloy-rpc-types.workspace = true
+
+# revm
+revm.workspace = true
+
+# stateless
+stateless-core = { path = "../stateless-core" }
+
+# misc
+bincode.workspace = true
+dashmap.workspace = true
+eyre.workspace = true
+lz4_flex.workspace = true
+redb.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+thiserror.workspace = true
+
+[dev-dependencies]
+tempfile.workspace = true
