+++ crates/stateless-db/src/cache.rs
@@ -1,0 +1,68 @@
+//! In-memory write-through cache for contract bytecodes.
+
+use std::{collections::HashMap, sync::Arc};
+
+use alloy_primitives::B256;
+use dashmap::DashMap;
+use revm::state::Bytecode;
+use stateless_core::db::{ContractStore, StoreResult};
+
+/// In-memory contract bytecode cache backed by persistent storage.
+///
+/// Reads check the in-memory `DashMap` first, falling back to the persistent store.
+/// Writes go to both memory and disk (write-through).
+pub struct ContractCache {
+    memory: DashMap<B256, Bytecode>,
+    store: Arc<dyn ContractStore>,
+}
+
+impl ContractCache {
+    /// Creates a new contract cache backed by the given persistent store.
+    pub fn new(store: Arc<dyn ContractStore>) -> Self {
+        Self { memory: DashMap::new(), store }
+    }
+
+    /// Retrieves contract bytecodes, checking memory first then persistent store.
+    ///
+    /// Returns `(found, missing)`.
+    pub fn get(&self, hashes: &[B256]) -> StoreResult<(HashMap<B256, Bytecode>, Vec<B256>)> {
+        let mut found = HashMap::new();
+        let mut not_in_memory = Vec::new();
+
+        for &hash in hashes {
+            if let Some(entry) = self.memory.get(&hash) {
+                found.insert(hash, entry.value().clone());
+            } else {
+                not_in_memory.push(hash);
+            }
+        }
+
+        if not_in_memory.is_empty() {
+            return Ok((found, Vec::new()));
+        }
+
+        let (from_disk, missing) = self.store.get_contracts(&not_in_memory)?;
+
+        for (hash, bytecode) in &from_disk {
+            self.memory.insert(*hash, bytecode.clone());
+        }
+        found.extend(from_disk);
+
+        Ok((found, missing))
+    }
+
+    /// Adds contract bytecodes to both memory cache and persistent store (write-through).
+    pub fn insert(&self, codes: &[(B256, Bytecode)]) -> StoreResult<()> {
+        if codes.is_empty() {
+            return Ok(());
+        }
+
+        self.store.add_contracts(codes)?;
+
+        for (hash, bytecode) in codes {
+            self.memory.insert(*hash, bytecode.clone());
+        }
+
+        Ok(())
+    }
+}
